# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Expense submission, approval, and payment processing system",
        "title": "Expense Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "tonny.wijayajuly2001@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/manager/expense-logs": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get paginated list of audit logs for expenses status changes (manager only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "Get audit logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by expense ID",
                        "name": "expense_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.ExpenseAuditLog"
                                    }
                                },
                                "meta": {
                                    "$ref": "#/definitions/controllers.PaginationMeta"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "constants.ExpenseStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected",
                "completed"
            ],
            "x-enum-varnames": [
                "ExpenseStatusPending",
                "ExpenseStatusApproved",
                "ExpenseStatusRejected",
                "ExpenseStatusCompleted"
            ]
        },
...
        "controllers.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "models.ExpenseAuditLog": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "expense_id": {
                    "type": "integer"
                },
                "from_status": {
                    "$ref": "#/definitions/constants.ExpenseStatus"
                },
                "id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "to_status": {
                    "$ref": "#/definitions/constants.ExpenseStatus"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}